refactor(db): Workroom capture and restore get their own module
The substrate ratchet caught this honestly: adding the restore planner took
backlog-recovery-bundle.ts to 900 lines and productionModuleHotspotCount from
63 to 64. Widening the ratchet would have been the wrong answer — the signal was
correct.

They are two concerns. backlog-recovery-bundle.ts owns the BUNDLE FORMAT for
epics and backlog items; workroom-recovery.ts owns the Workroom rows that say
where the work actually is. The first re-exports the second, so every existing
importer keeps working and there is still one place to look.

686 + 225 lines, substrate guard passes, 39 guards pass, 18 tests pass.
