+++ packages/db/package.json
@@ -59,6 +59,7 @@
     "./cada-readiness": "./src/cada-readiness.ts",
     "./regulation-applicability": "./src/regulation-applicability.ts",
     "./reference-model-applicability": "./src/reference-model-applicability.ts",
+    "./workroom-recovery": "./src/workroom-recovery.ts",
     "./seed-deliberation": "./src/seed-deliberation.ts",
     "./edge-node-types": "./src/edge-node-types.ts",
     "./wiki-taxonomy": "./src/wiki-taxonomy.ts",
